+++ src/noise/mod.rs
@@ -130,7 +130,7 @@ where
 /// Represents period at which the noise may repeat, the inverse of frequency.
 /// You can think of this like the scale of the noise.
 #[derive(Debug, Clone, Copy, PartialEq)]
-pub struct Period(pub f64);
+pub struct Period(pub f32);
 
 /// This describes typical settings for a [`NoiseOp`] that creates noise based purely on space and
 /// seeds.
@@ -143,7 +143,7 @@ pub struct SpatialNoiseSettings {
 impl SpatialNoiseSettings {
     /// Constructs a new [`SpatialNoiseSettings`] based on the given seed and
     /// [`period`](SpatialNoiseSettings::period).
-    pub fn new(seed: u64, period: f64) -> Self {
+    pub fn new(seed: u64, period: f32) -> Self {
         Self {
             period: Period(period),
             rng: NoiseRng::new_seed(seed),
